feat: Consultoria - alterei alguns arquivos para typescript - adicionei o .editorconfig. - ...
