+++ package.json
@@ -17,6 +17,7 @@
   "devDependencies": {
     "@types/react": "^18.0.26",
     "@types/react-dom": "^18.0.9",
+    "@types/styled-components": "^5.1.26",
     "@vitejs/plugin-react": "^3.0.0",
     "vite": "^4.0.0"
   }
