+++ src/components/Card/style.js
@@ -1,11 +1,11 @@
 import styled from "styled-components";
 
 const colors = {
-    green: '#e2ffef',
-    red: '#ffe2e2',
-    yellow: '#ffffe2',
-    blue: '#e2f1ff',
-    purple: '#f0e6ff'
+  green: '#e2ffef',
+  red: '#ffe2e2',
+  yellow: '#ffffe2',
+  blue: '#e2f1ff',
+  purple: '#f0e6ff'
 }
 
 export const Container = styled.div`
