+++ src/helpers/consts.ts
@@ -1,0 +1,1 @@
+export const removeNullSpaces = /\s/g
