--- src/helpers/getCardsFromLocalStorage.js
@@ -1,7 +1,0 @@
-export const getCardsFromLocalStorage = () => {
-    const cards = localStorage.getItem('cards');
-    if (cards) {
-      return JSON.parse(cards);
-    }
-    return cards;
-}
