+++ src/styles/Theme.js
@@ -1,12 +1,13 @@
 export const lightTheme = {
-    body: '#FFF',
-    text: '#333',
-    inputText: '#333',
-    inputHover: '#ddd',
+  body: '#FFF',
+  text: '#333',
+  inputText: '#333',
+  inputHover: '#ddd',
 }
+
 export const darkTheme = {
-    body: '#161616',
-    text: '#333',
-    inputText: '#333',
-    inputHover: '#ddd',
+  body: '#161616',
+  text: '#333',
+  inputText: '#333',
+  inputHover: '#ddd',
 }
